Moved HTTP Requests interactions related code to a separate Service
